+++ app/src/main/res/layout/fragment_movie_detail.xml
@@ -121,34 +121,6 @@
                 android:orientation="vertical">
             </LinearLayout>
 
-            <LinearLayout
-                android:id="@+id/trailer_area"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:clickable="true"
-                android:orientation="horizontal"
-                android:paddingLeft="20dp">
-
-                <ImageButton
-                    android:id="@+id/start_video_button"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginBottom="8dp"
-                    android:background="@android:color/white"
-                    android:src="@android:drawable/ic_media_play"/>
-
-                <TextView
-                    android:id="@+id/trailer_text_view"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:paddingLeft="8dp"
-                    android:paddingStart="8dp"
-                    android:text="@string/no_trailers_avail"
-                    tools:text="Trailer One">
-
-                </TextView>
-
-            </LinearLayout>
 
             <View
                 android:layout_width="wrap_content"
