+++ app/src/main/res/layout/list_item_trailer.xml
@@ -1,5 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
-<linearlayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
               xmlns:tools="http://schemas.android.com/tools"
               android:layout_width="wrap_content"
               android:layout_height="wrap_content"
@@ -8,6 +7,8 @@
 
     <ImageButton
         android:id="@+id/play_button"
+        android:layout_width="40dp"
+        android:layout_height="40dp"
         android:background="@android:color/holo_green_dark"
         android:scaleType="center"
         android:src="@android:drawable/ic_media_play"
@@ -19,4 +20,4 @@
         android:layout_height="wrap_content"
         tools:text="Trailer 1"/>
 
-</linearlayout>
+</LinearLayout>
